Implements `node_inputs_resolver`.
`node_inputs_resolver` is responsible for assembling each InputSpec. Normally it is simply a dict merge, but if there is an input_graph_ref to the ARTIFACT_MULTIMAP_LIST, then node_inputs_resolver also takes care for the zip or cartesian product of the result.

Since Each InputSpec can have dependency on another, we do the topological sort (error on cycle) in advance. This general ability of InputSpec referring to another InputSpec is crucial to correctly translates BaseChannel semantics to the InputSpec IR.

PiperOrigin-RevId: 458411999